refactor: move apis out from util
